update user type

+++ models/user.go
@@ -15,7 +15,7 @@ User {
 type User struct {
 	Uid     int             `json:"Uid"`
 	Name    string          `json:"Name"`
-	Balance int             `json:"Balance"`
+	Balance float64         `json:"Balance"`
 	Owes    map[int]float64 `json:"Owes"`
 	Owed    map[int]float64 `json:"Owed"`
 	Number  string          `json:"Number"`
